Fix Autolog timing to not DDOS server.
This is because APOS can't run a loop with the way Login Listener is written, this will start at 5s(+=1s) autolog for 2 attempts, then 15s(+=5s) autolog for the next 8 attempts (2 minute total)  (there is currently only a 5s autolog, no WONDER there are issues when server reboots. ) then it will start at around a 1min login delay and ramp up to ~2.7 min delay at n = 50, and a ~ 4.2 min delay as n => 500. and a 4.6 min delay as n> 500k (so no worries about leaving a client autologging, it will converge on ~4.6666 (6 repeating) mins at infinity).

Toggling autologin will not reset your autologin counter, but when done the bot controller will always immediately attempt to log in, and if it cannot then it will go back to waiting. Perhaps a future addition can reset the counter? 

system.out.println shows exactly how long till next autolog attempt, shown in seconds, as well as how many auto login attempts (since you were last online, this resets when you login)

1 small bug, onLoginResponse appears to process 2 times before bot goes back to

 `if (System.currentTimeMillis() < next_attempt) return;` to wait for next_attempt system time. 

This does produce 2x chat messages in the client console, as well as counts "i" twice as fast as intender. HOWEVER, the bot only logs in 1x per cycle, and it still produces the intended login timing results when tested (5s, 5s, 15s, 15s, ... , 30s, ... , 60s, ... ,120s)

this calculator can be used to visualize the different equations as "i" increases in value. 
https://www.desmos.com/calculator

Graphs are an interesting way to visualize it. The value of the number x in 
x = (i*15)/(i+60)
will begin to converge on a fixed value as "i" approaches ~240, notice we don't get too much increase in the beginning, that is good. Then we apply that x(i) to the equation y = (Math.random() * 30000) + (30000x(i)) + 30000. we want large values of i to converge on y = 5 for about a 4 - 5 min login retry timer after ~150-200 retrys. The size of (60) determines how flat our curve is initially (it has a reducing effect on the growth of the numerator). The Size of 15 determines where on the y axis our curve will converge to. convergence doesn't matter too much because we are only going to i=240, so we wont check very large values. All said and done we get this result (calculated with excel):