1 frame screen refresh for gui elements
Briefly toggles on render for 1 frame (20ms) every 60s that pass while render is off, this is to refresh onscreen gui with new information, as well as prevent leaving render off longterm defaulting to the zoomed out minimap image that is common. The call for render toggle is on game message, so this only works while botting (that is ideal) and will trigger if time is right and a game message appears. Then onPaintTick is active at that point, so we let that call the 1 frame toggle off.  Using game message now could take a while to get another game message.

Stormy said this should not affect the bot resource usage at all really.